python/voice-ai-pipecat/requirements.txt
pipecat-ai[websocket,google,whisper,piper,silero,local]<1.0
faster-whisper
piper-tts
pyaudio
fastapi
uvicorn[standard]
websockets
python-dotenv
# pipecat-ai 0.0.108+ no longer pulls nltk as a transitive dep; the Dockerfile
# pre-downloads punkt_tab for Piper TTS sentence tokenization, so add it here.
nltk
